made backButton a component
